Introduce the AbstractSaver class
As saving notes simply to the .txt file seems to be not the only
opportunity in the near future thus the AbsatractSaver class is
introduced to make use of inheritance conception. The NoteSaver class
was renamed to NoteSaverTxt and refactored the way it is the child of
the AbstactSaver class. Also got rid of check_note_is_valid_type()
method as it goes against duck-typing idiom.